Move rotating text into it's own component
